Four things nothing used, and a cooker that halved every index it wrote
Ten engine libraries are unreachable from both app binaries. Most have a
credible path to a user and should wait. These four do not, and carrying them
costs more than the git history that keeps them.

//engine/net, 1,572 source lines and a 643-line test, zero references anywhere
outside its own package. Its real socket has never been executed by anything,
and its prerequisites -- a stable networked identity for objects the app
addresses by vector index, a deterministic simulation, a server binary -- are
each longer than the module. It is good work with no consumer and no near path
to one.

//engine/platform:audio_file, 151 lines, zero references and zero tests. It is
what made README.md's opening paragraph false: the engine does not decode
compressed audio, because nothing calls the thing that would.

THE STEREO PATH. Bit-identical per eye at 0.075 ms against 0.103 for two passes,
and there is no output device for it in this repository and no plan for one. It
was not free: Shading::LitStereo, a GeometryPass mode, a pipeline variant built
for every opaque lit material, a second vertex stage in lit.metal, and two
fields on the per-draw uniform block that every draw in every pass paid for.
FrameUniforms goes 656 -> 576 bytes.

THE SKINNING DISPATCH IN apps/world, which produced a buffer the app never read.
SkinToBuffers' output reaches the world only through Renderer::PosedVertices,
whose only consumer is BuildSceneAccel, which apps/world does not call -- the
forward path poses the mesh in its own vertex stage. One compute encoder and one
palette upload per frame, for nothing, and it is visible in the timeline now
that there is one. The engine keeps the entry point; ray-traced shadows from a
posed mesh need it and tests/skinned covers that.

AND THE PACK COOKER WAS WRITING BROKEN MESHES. Mesh::indices is
std::vector<std::uint32_t>. AddMesh set index_stride to sizeof(uint16_t), sized
the index block at half what the indices need, and memcpy'd that many bytes out
of the uint32 array -- so a cooked mesh held the low half of its index list
reinterpreted as pairs of shorts, and the reader's stride check agreed with the
writer so nothing rejected it.

It survived because the test could not fail. pack_test compared
`sphere.indices.size() * sizeof(std::uint16_t)` bytes -- the same wrong number --
under a comment reading "BYTE FOR BYTE, which is the whole claim". The
assertion is the LENGTH now, taken from the span the reader returns rather than
recomputed by the test, and it is demonstrated rather than asserted: restoring
the 16-bit stride fails two checks.

  before: "and index data identical too"  ok      (comparing half the data)
  after : the same test against the old writer, 2 checks FAIL

README corrected: the AudioToolbox claim, the stereo and networking sections,
65,000 lines and 69 test targets. Three claims that were true of the test suite
and not of the program are now labelled as such -- the GPU-driven figure is a
tests/manyobjects scene of 6000 copies of one mesh, and the demo valley's 196
mostly-single-instance batches behave differently and are given their own
numbers; the asset package and hot reload are "a pipeline with no ends rather
than a pipeline"; and there is no runtime content pipeline at all, because every
mesh, texture and sound in apps/world is generated by C++ at startup.

69 tests pass, from 71. Screenshots byte-identical excluding the HUD.

+++ engine/asset/pack.cc
@@ -104,14 +104,24 @@ void PackWriter::AddMesh(std::string_view name, const Mesh& mesh) {
     h.vertex_count = std::uint32_t(mesh.vertices.size());
     h.index_count = std::uint32_t(mesh.indices.size());
     h.vertex_stride = std::uint32_t(sizeof(VertexIn));
-    h.index_stride = std::uint32_t(sizeof(std::uint16_t));
+    // 32-BIT, which is what Mesh::indices actually holds. It said uint16_t --
+    // so the header advertised a stride the data did not have, the index block
+    // was sized at HALF what the indices need, and the memcpy below copied that
+    // many bytes out of a uint32 array. What came back was the low half of the
+    // index list reinterpreted as pairs of shorts. It survived because
+    // pack_test compared the same halved length, under a comment reading
+    // "BYTE FOR BYTE, which is the whole claim" -- a test that cannot fail.
+    //
+    // 32-bit indices are not an option here either: engine/render uploads
+    // uint32 and the README's "a mesh has no vertex ceiling" depends on it.
+    h.index_stride = std::uint32_t(sizeof(std::uint32_t));
     h.centre[0] = mesh.bounds.center.x;
     h.centre[1] = mesh.bounds.center.y;
     h.centre[2] = mesh.bounds.center.z;
     h.radius = mesh.bounds.radius;
 
     const std::size_t vertex_bytes = mesh.vertices.size() * sizeof(VertexIn);
-    const std::size_t index_bytes = mesh.indices.size() * sizeof(std::uint16_t);
+    const std::size_t index_bytes = mesh.indices.size() * sizeof(std::uint32_t);
     // The vertex block starts on a sixteen-byte boundary WITHIN the payload, so
     // that an aligned payload gives an aligned vertex block.
     const std::size_t vertex_at = Align16(sizeof(MeshHeader));
@@ -323,7 +333,7 @@ bool Pack::GetMesh(std::string_view name, CookedMesh* out) const {
     // layout -- and a mesh reinterpreted at the wrong stride is not a crash,
     // it is a shape nobody recognises, which reads as a broken exporter.
     if (h.vertex_stride != sizeof(VertexIn) ||
-        h.index_stride != sizeof(std::uint16_t))
+        h.index_stride != sizeof(std::uint32_t))
         return false;
 
     const std::size_t vertex_at = Align16(sizeof(MeshHeader));
